finished wood per sec display + few improvements on UI

+++ Motor2D/j1EntityController.h
@@ -87,7 +87,7 @@ public:
 	bool building = false;
 	int death_time = 0;
 	int mill_max_villagers = 0;
-	float worker_wood_production = 0.0f;
+	int worker_wood_production = 0;
 	int test_counter = 0;
 	
 	buildingType structure_beingbuilt = NONE_BUILDING;
